LayoutTests/platform/blackberry/fast/filesystem/file-writer-write-overlapped-expected.txt

CONSOLE MESSAGE: line 163: ReferenceError: Can't find variable: webkitRequestFileSystem
FAIL: Timed out waiting for notifyDone to be called
Test using FileWriter.seek to write overlapping existing data in a file.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".
